refactor: :label: update import paths for database and schema modules
