+++ app/(Auth)/signin/google/callback/route.ts
@@ -2,9 +2,9 @@ import { GoogleUser } from "@/lib/types/auth.types";
 import { generateIdFromEntropySize } from "lucia";
 import { GoogleTokens, OAuth2RequestError } from "arctic";
 import { cookies } from "next/headers";
-import db from "@/db/drizzle";
+import db from "@/lib/db/drizzle";
 import { eq } from "drizzle-orm";
-import { userTable } from "@/db/db.schema";
+import { userTable } from "@/lib/db/db.schema";
 import { NextResponse } from "next/server";
 import { google, lucia } from "@/lib/apis/auth.api";
