+++ lib/db/drizzle.ts
@@ -2,7 +2,7 @@ import "dotenv/config";
 
 import { neon } from "@neondatabase/serverless";
 import { drizzle } from "drizzle-orm/neon-http";
-import * as schema from "@/db/db.schema";
+import * as schema from "./db.schema";
 
 export const sql = neon(process.env.NEON_DATABASE_URL!);
 const db = drizzle(sql, { logger: false, schema });
