+++ lib/db/migrate.ts
@@ -1,6 +1,6 @@
 import "dotenv/config";
 import { migrate } from "drizzle-orm/neon-http/migrator";
-import db from "@/db/drizzle";
+import db from "./drizzle";
 
 // This will run migrations on the database, skipping the ones already applied
