+++ lib/utils/drizzle.utils.ts
@@ -1,4 +1,4 @@
-import { courseTable } from "@/db/db.schema";
+import { courseTable } from "@/lib/db/db.schema";
 import { CoursesFilter } from "@/lib/types/manage.courses.types";
 import { eq, gte, lte, lt, gt, and } from "drizzle-orm";
 
@@ -7,13 +7,23 @@ export const coursesFilter = (filter: CoursesFilter) => {
     return eq(courseTable.draftMode, false);
   }
   if (filter === "ongoing") {
-    return and(gte(courseTable.endDate, new Date()), lte(courseTable.startDate, new Date()), eq(courseTable.draftMode, false));
+    return and(
+      gte(courseTable.endDate, new Date()),
+      lte(courseTable.startDate, new Date()),
+      eq(courseTable.draftMode, false),
+    );
   }
   if (filter === "archived") {
-    return and(lt(courseTable.endDate, new Date()), eq(courseTable.draftMode, false));
+    return and(
+      lt(courseTable.endDate, new Date()),
+      eq(courseTable.draftMode, false),
+    );
   }
   if (filter === "starting soon") {
-    return and(gt(courseTable.startDate, new Date()), eq(courseTable.draftMode, false));
+    return and(
+      gt(courseTable.startDate, new Date()),
+      eq(courseTable.draftMode, false),
+    );
   }
   if (filter === "draft") {
     return eq(courseTable.draftMode, true);
